move all joint-specific patterns out of cnf and into joint
